perf: update the documentation and finalize the plumber version
BREAKING CHANGE: **This repository contains breaking changes with older versions that are built on
top of `s6-overlay`!**

Sorry for any inconvenience caused. It was a decision between making some breaking changes to expand
this container or just abandoning this one and creating a new one.

Since my use case now also requires creating bridge networks to access a local network, I had to
introduce a new way of configuring this container. This renders the older versions intentionally
incompatible for any user to see that something is wrong and hopefully visit here again to easily
fix it.